Always show open in another tab link for legends that have url
